\name{cwhprint-package}
\alias{cwhprint-package}
\docType{package}
\title{
cwh's printing functions
}
\description{
Miscellaneous functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhprint\cr
Version: \tab 4.3.0\cr
Date: \tab 2006-02-24\cr
Depends: \tab R (>= 0.99), cwhstring\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann\cr
}

Index:
\preformatted{
cwhprint-package        cwh's print functions
heading                 Write a line of text with underlining and blank
                        lines
printP                  Print without square brackets, expression
                        values together with their call strings
tex.table               Convert a data matrix into LaTeX code.
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
